+++ app/build.gradle
@@ -51,7 +51,9 @@ dependencies {
     implementation 'com.android.volley:volley:1.1.1'
     implementation "androidx.cardview:cardview:1.0.0"
     implementation "androidx.recyclerview:recyclerview:1.1.0"
-    implementation 'com.github.kittinunf.fuel:fuel:2.2.0'
-
-
+//    implementation 'com.github.kittinunf.fuel:fuel:2.2.0'
+    implementation 'com.github.kittinunf.fuel:fuel-json:2.2.0'
+    implementation 'com.tomatobean:jsonparser:1.0.10'
+    implementation("org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.20.0")
+    implementation 'com.github.chintan369:Material-Event-On-Calendar:1.0.8'
 }
