+++ app/src/main/res/layout/calender_month.xml
@@ -10,11 +10,9 @@
     android:background="@drawable/button_background"
     >
 
-    <CalendarView
-        android:id="@+id/calendarView"
-        android:layout_centerVertical="true"
+    <com.chintanpatel.materialeventcalendar.CalenderView
+        android:id="@+id/eventCalendar"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-         />
+        android:layout_height="match_parent"/>
 </RelativeLayout>
 </RelativeLayout>
